Add AdminPage, CreateProduct, Modified HomePage and Login
The admin page was added, where you can view the products and create, delete and modify them. The page to create a product was added, with its respective image. The homepage was modified to display the images of the created products and validation and password policy were added to the login.